Added Driver
I have added part of the driver code so that the program can be made
to eventually choose a song. Currently, it will load selfie.png in
the project directory and use that as the basis for selecting a
song. Soon, we will want to use Jacob's code to take a picture, move
it to selfie.png, and use the image to choose a song.